+++ servers/noMulter/api/users/_userId@number/$relay.ts
@@ -1,11 +1,11 @@
 /* eslint-disable */
-import { Express, RequestHandler } from 'express'
-import { Schema } from 'fast-json-stringify'
-import { HttpStatusOk } from 'aspida'
-import { Deps, depend } from 'velona'
-import { ServerMethods } from '../../../$server'
-import { AdditionalRequest } from './../hooks'
-import { Methods } from './'
+import type { Express, RequestHandler } from 'express'
+import type { Schema } from 'fast-json-stringify'
+import type { HttpStatusOk } from 'aspida'
+import type { ServerMethods } from '../../../$server'
+import type { AdditionalRequest } from './../hooks'
+import type { Methods } from './'
+import { Injectable, depend } from 'velona'
 
 type AddedRequestHandler = RequestHandler extends (req: infer U, ...args: infer V) => infer W ? (req: U & Partial<AdditionalRequest>, ...args: V) => W : never
 type Hooks = {
@@ -25,13 +25,13 @@ export function defineResponseSchema<T extends { [U in keyof ControllerMethods]?
 }
 
 export function defineHooks<T extends Hooks>(hooks: (app: Express) => T): (app: Express) => T
-export function defineHooks<T extends Record<string, any>, U extends Hooks>(deps: T, cb: (d: Deps<T>, app: Express) => U): { (app: Express): U; inject(d: Deps<T>): (app: Express) => U }
-export function defineHooks<T extends Record<string, any>>(hooks: (app: Express) => Hooks | T, cb?: (deps: Deps<T>, app: Express) => Hooks) {
+export function defineHooks<T extends Record<string, any>, U extends Hooks>(deps: T, cb: (d: T, app: Express) => U): Injectable<T, [Express], U>
+export function defineHooks<T extends Record<string, any>>(hooks: (app: Express) => Hooks | T, cb?: (deps: T, app: Express) => Hooks) {
   return cb && typeof hooks !== 'function' ? depend(hooks, cb) : hooks
 }
 
 export function defineController(methods: (app: Express) => ControllerMethods): (app: Express) => ControllerMethods
-export function defineController<T extends Record<string, any>>(deps: T, cb: (d: Deps<T>, app: Express) => ControllerMethods): { (app: Express): ControllerMethods; inject(d: Deps<T>): (app: Express) => ControllerMethods }
-export function defineController<T extends Record<string, any>>(methods: (app: Express) => ControllerMethods | T, cb?: (deps: Deps<T>, app: Express) => ControllerMethods) {
+export function defineController<T extends Record<string, any>>(deps: T, cb: (d: T, app: Express) => ControllerMethods): Injectable<T, [Express], ControllerMethods>
+export function defineController<T extends Record<string, any>>(methods: (app: Express) => ControllerMethods | T, cb?: (deps: T, app: Express) => ControllerMethods) {
   return cb && typeof methods !== 'function' ? depend(methods, cb) : methods
 }