+++ servers/noTypedParams/$server.ts
@@ -1,6 +1,6 @@
 /* eslint-disable */
+import type { LowerHttpMethod, AspidaMethods, HttpStatusOk, AspidaMethodParams } from 'aspida'
 import path from 'path'
-import { LowerHttpMethod, AspidaMethods, HttpStatusOk, AspidaMethodParams } from 'aspida'
 import express, { Express, RequestHandler, Request } from 'express'
 import multer, { Options } from 'multer'
 import { validateOrReject, ValidatorOptions } from 'class-validator'
@@ -80,7 +80,7 @@ const parseJSONBoby: RequestHandler = (req, res, next) => {
 }
 
 const createValidateHandler = (validators: (req: Request) => (Promise<void> | null)[]): RequestHandler =>
-  (req, res, next) => Promise.all(validators(req)).then(() => next()).catch(() => res.sendStatus(400))
+  (req, res, next) => Promise.all(validators(req)).then(() => next()).catch(err => res.status(400).send(err))
 
 const formatMulterData = (arrayTypeKeys: [string, boolean][]): RequestHandler => ({ body, files }, _res, next) => {
   for (const [key] of arrayTypeKeys) {
@@ -143,6 +143,7 @@ const asyncMethodToHandler = (
 
 export default (app: Express, options: FrourioOptions = {}) => {
   const basePath = options.basePath ?? ''
+  const validatorOptions: ValidatorOptions = { validationError: { target: false }, ...options.validator }
   const hooks0 = hooksFn0(app)
   const hooks1 = hooksFn1(app)
   const ctrlHooks0 = ctrlHooksFn0(app)
@@ -159,7 +160,7 @@ export default (app: Express, options: FrourioOptions = {}) => {
     hooks0.onRequest,
     ctrlHooks0.onRequest,
     createValidateHandler(req => [
-      Object.keys(req.query).length ? validateOrReject(Object.assign(new Validators.Query(), req.query), options.validator) : null
+      Object.keys(req.query).length ? validateOrReject(Object.assign(new Validators.Query(), req.query), validatorOptions) : null
     ]),
     asyncMethodToHandler(controller0.get)
   ])
@@ -170,8 +171,8 @@ export default (app: Express, options: FrourioOptions = {}) => {
     uploader,
     formatMulterData([]),
     createValidateHandler(req => [
-      validateOrReject(Object.assign(new Validators.Query(), req.query), options.validator),
-      validateOrReject(Object.assign(new Validators.Body(), req.body), options.validator)
+      validateOrReject(Object.assign(new Validators.Query(), req.query), validatorOptions),
+      validateOrReject(Object.assign(new Validators.Body(), req.body), validatorOptions)
     ]),
     methodToHandler(controller0.post)
   ])
@@ -186,7 +187,7 @@ export default (app: Express, options: FrourioOptions = {}) => {
     uploader,
     formatMulterData([['empty', false], ['vals', false], ['files', false]]),
     createValidateHandler(req => [
-      validateOrReject(Object.assign(new Validators.MultiForm(), req.body), options.validator)
+      validateOrReject(Object.assign(new Validators.MultiForm(), req.body), validatorOptions)
     ]),
     methodToHandler(controller2.post)
   ])
@@ -219,7 +220,7 @@ export default (app: Express, options: FrourioOptions = {}) => {
     hooks1.onRequest,
     parseJSONBoby,
     createValidateHandler(req => [
-      validateOrReject(Object.assign(new Validators.UserInfo(), req.body), options.validator)
+      validateOrReject(Object.assign(new Validators.UserInfo(), req.body), validatorOptions)
     ]),
     ...ctrlHooks1.preHandler,
     methodToHandler(controller5.post)